# [Webportal] static content caching optimization