--- a/src/webportal/server/config/express.js
+++ b/src/webportal/server/config/express.js
@@ -43,9 +43,18 @@ app.use(favicon(path.join(appRoot.path, 'dist', 'favicon.ico')));
 // setup the root path
 const oneMonth = 3600 * 24 * 30 * 1000;
 
-app.use('/assets', express.static(path.join(appRoot.path, 'dist/assets'), { maxAge: oneMonth }));
-app.use('/styles', express.static(path.join(appRoot.path, 'dist/styles'), { maxAge: oneMonth }));
-app.use('/scripts', express.static(path.join(appRoot.path, 'dist/scripts'), { maxAge: oneMonth }));
+app.use(
+  '/assets',
+  express.static(path.join(appRoot.path, 'dist/assets'), { maxAge: oneMonth }),
+);
+app.use(
+  '/styles',
+  express.static(path.join(appRoot.path, 'dist/styles'), { maxAge: oneMonth }),
+);
+app.use(
+  '/scripts',
+  express.static(path.join(appRoot.path, 'dist/scripts'), { maxAge: oneMonth }),
+);
 app.use(express.static(path.join(appRoot.path, 'dist')));
 
 // catch 404 and forward to error handler