--- a/src/webportal/src/app/home/index.ejs
+++ b/src/webportal/src/app/home/index.ejs
@@ -9,6 +9,6 @@
 </head>
 <body>
   <div id="content"></div>
-  <script src="/scripts/env.js"></script>
+  <script src="/env.js"></script>
 </body>
 </html>