--- a/src/webportal/src/app/layout/layout.component.ejs
+++ b/src/webportal/src/app/layout/layout.component.ejs
@@ -135,6 +135,6 @@
 
   </div>
   <script>window.PAI_VERSION = <%=JSON.stringify(htmlWebpackPlugin.options.version)%></script>
-  <script src="/scripts/env.js"></script>
+  <script src="/env.js"></script>
 </body>
 </html>