There are 2 main issues about static content (js, style css, font, svg, .., etc) in current webportal:

  - They are caching-disabled. The `max-age` in the response header for them is 0. Thus, the browser will always download these contents, which increases both the front-end loading time and the server's traffic.
  - Many js files are splitted into very small chunks (<1kb), which means they should also be transferred by new HTTP connection.

This PR sets a `max-age` for those contents,  use `contenthash` in webpack to ensure correct caching, 
and only js files larger than 30kb would be chunked.

Loading `submit.html` before and after optimization (tested for 5 times in each case):
   - Before optimization + First loading: 4.19s ~ 8.61s
   - Before optimization + Second loading: 4.21s~6.49s
   - After optimization   + First loading: 3.71s ~ 4.76s
   - After optimization   + Second loading: **0.65s~0.94s**
 
Since "second loading" would be the main use case, the user experience could be enhanced a lot by this optimization.

Chrome `Network` page for "Before optimization + Second loading":

![before](https://user-images.githubusercontent.com/7499023/68919441-1b7af100-07ad-11ea-8fc5-3d56decb2766.png)

Chrome `Network` page for "After optimization + Second loading":

![after](https://user-images.githubusercontent.com/7499023/68919472-351c3880-07ad-11ea-808a-3266e4873d9e.png)